evalutation is now an integer

+++ Chess-Challenge/src/My Bot/MyBot.cs
@@ -9,12 +9,12 @@ public class MyBot : IChessBot {
         bool white = board.IsWhiteToMove;
         Move[] moves = board.GetLegalMoves();
         Move move = moves[new Random().Next(moves.Length)];
-        float boardEval = Evaluation(board); // get board evaluation before the bot makes a move
+        int boardEval = Evaluation(board); // get board evaluation before the bot makes a move
 
         foreach (Move possibleMove in moves) {
             // play possibleMove, then check if that move improves the bot's eval. if it does, set move = possibleMove
             board.MakeMove(possibleMove);
-            float eval = Evaluation(board);
+            int eval = Evaluation(board);
             if (white && eval > boardEval) {
                 boardEval = eval;
                 move = possibleMove;
@@ -28,9 +28,11 @@ public class MyBot : IChessBot {
         return move;
     }
 
-    float Evaluation(Board board) {
+
+
+    int Evaluation(Board board) {
         if (board.IsInCheckmate()) return board.IsWhiteToMove ? -10000 : 10000; // if its checkmate, dont do any of the following checks
-        float evaluation = 0;
+        int evaluation = 0;
         for (int i = 0; i < 64; i++) { // go through all 64 squares
             int pieceType = (int) GetPiece(board, i).PieceType; // get piece type in integer form (0 for nothing, 1 for pawn, etc.)
             int value = pieceValues[pieceType]; // get the value of the piece at the square using the pieceValues array
